+++ Projects/App/Project.swift
@@ -13,7 +13,7 @@ let project = Project.makeModule(
       .core(.sSRoot),
       .core(.coreLayers),
       .feature(.sent),
-      .feature(.inventory)
+      .feature(.inventory),
     ],
     testDependencies: [],
     infoPlist: [
