+++ Projects/Feature/Inventory/Sources/InventoryBuilderView/InventoryBuilderView.swift
@@ -10,11 +10,8 @@ import SwiftUI
 
 public struct InventoryBuilderView: View {
   public init() {}
-  
+
   public var body: some View {
-    HStack {
-      
-    }
+    HStack {}
   }
-  
 }
