+++ Projects/Feature/Sent/Sources/SentEnvelopeFilter/SentEnvelopeFilter.swift
@@ -20,6 +20,9 @@ struct SentEnvelopeFilter {
     var isOnAppear = false
     @Shared var textFieldText: String
     var sentPeopleAdaptor: SentPeopleAdaptor
+
+    // MARK: - Scope
+
     var header: HeaderViewFeature.State = .init(.init(title: "필터", type: .depth2Default))
     var sliderProperty: CustomSlider = .init(start: 0, end: 100_000, width: UIScreen.main.bounds.size.width - 42)
     var customTextField: CustomTextField.State
