+++ Projects/Feature/Sent/Sources/SentMainView/SentMainView.swift
@@ -60,22 +60,25 @@ struct SentMainView: View {
       )) {
         store.send(.view(.setFilterDialSheet(true)))
       }
-      ZStack {
-        // TODO: Navigation 변경
-        NavigationLink(state: SentRouter.Path.State.sentEnvelopeFilter(.init(sentPeople: [
-          .init(name: "춘자"),
-          .init(name: "복자"),
-          .init(name: "흑자"),
-          .init(name: "헬자"),
-          .init(name: "함자"),
-          .init(name: "귀자"),
-          .init(name: "사귀자"),
-        ]))) {
-          SSButton(Constants.notSelectedFilterButtonProperty) {
-            store.send(.view(.filterButtonTapped))
-          }
-          .allowsHitTesting(false)
-        }
+      SSButton(Constants.notSelectedFilterButtonProperty) {
+        store.send(.view(.filterButtonTapped))
+        //      ZStack {
+        //        // TODO: Navigation 변경
+        //        NavigationLink(state: SentRouter.Path.State.sentEnvelopeFilter(.init(sentPeople: [
+        //          .init(name: "춘자"),
+        //          .init(name: "복자"),
+        //          .init(name: "흑자"),
+        //          .init(name: "헬자"),
+        //          .init(name: "함자"),
+        //          .init(name: "귀자"),
+        //          .init(name: "사귀자"),
+        //        ]))) {
+        //          SSButton(Constants.notSelectedFilterButtonProperty) {
+        //            store.send(.view(.filterButtonTapped))
+        //          }
+        //          .allowsHitTesting(false)
+        //        }
+        //      }
       }
     }
     .frame(maxWidth: .infinity, alignment: .topLeading)
