+++ Projects/Feature/Sent/Sources/SentRouterView/SentRouter.swift
@@ -32,8 +32,21 @@ struct SentRouter {
         switch action {
         case .element(id: _, action: .sentEnvelopeFilter):
           return .none
+
         case .element(id: _, action: .sentMain(.delegate(.pushSearchEnvelope))):
-//          state.path.append(.searchEnvelope(SearchEnvelope.State()))
+          state.path.append(.searchEnvelope(SearchEnvelope.State()))
+          return .none
+
+        case .element(id: _, action: .sentMain(.delegate(.pushFilter))):
+          state.path.append(.sentEnvelopeFilter(.init(sentPeople: [
+            .init(name: "춘자"),
+            .init(name: "복자"),
+            .init(name: "흑자"),
+            .init(name: "헬자"),
+            .init(name: "함자"),
+            .init(name: "귀자"),
+            .init(name: "사귀자"),
+          ])))
           return .none
         default:
           return .none
